requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
fastapi>=0.104
uvicorn>=0.24
mlflow>=2.9
dvc>=3.30
deepchecks>=0.18
python-jose[cryptography]>=3.3
slowapi>=0.1.9
redis>=5.0
beautifulsoup4>=4.12
pyarrow>=14.0
scipy>=1.11
torch>=2.1
lxml>=4.9
joblib>=1.3
